+++ src/oldnews/providers/main.py
@@ -27,6 +27,7 @@ from ..commands import (
     Previous,
     PreviousUnread,
     RefreshFromTheOldReader,
+    RenameSubscription,
     ToggleShowAll,
 )
 
@@ -53,6 +54,7 @@ class MainCommands(CommandsProvider):
         yield from self.maybe(CopyFeedToClipboard)
         yield from self.maybe(CopyArticleToClipboard)
         yield from self.maybe(Copy)
+        yield from self.maybe(RenameSubscription)
         yield AddSubscription()
         yield ToggleShowAll()
         yield RefreshFromTheOldReader()