+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # OldNews ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Fixed subscriptions not in folders not showing in navigation.
+  ([#49](https://github.com/davep/oldnews/pull/49))
+
 ## v0.4.0
 
 **Released: 2026-01-20**