+++ ChangeLog.md
@@ -9,6 +9,7 @@
 - Fixed remotely-removed subscriptions appearing back in navigation.
   ([#50](https://github.com/davep/oldnews/pull/50))
 - Added the `AddSubscription` command for adding a new subscription feed.
+  ([#53](https://github.com/davep/oldnews/pull/53))
 
 ## v0.4.0