+++ ChangeLog.md
@@ -13,6 +13,7 @@
 - Fixed freshly-added subscriptions not having their article backlog
   downloaded. ([#55](https://github.com/davep/oldnews/pull/55))
 - Fixed unread count missing unread articles not in folders.
+  ([#56](https://github.com/davep/oldnews/pull/56))
 
 ## v0.4.0