+++ ChangeLog.md
@@ -12,6 +12,8 @@
   ([#53](https://github.com/davep/oldnews/pull/53))
 - Fixed freshly-added subscriptions not having their article backlog
   downloaded. ([#55](https://github.com/davep/oldnews/pull/55))
+- Fixed unread count missing unread articles not in folders.
+  ([#56](https://github.com/davep/oldnews/pull/56))
 
 ## v0.4.0