+++ ChangeLog.md
@@ -14,6 +14,8 @@
   downloaded. ([#55](https://github.com/davep/oldnews/pull/55))
 - Fixed unread count missing unread articles not in folders.
   ([#56](https://github.com/davep/oldnews/pull/56))
+- Added the `RenameSubscription` command for renaming the current
+  subscription. ([#59](https://github.com/davep/oldnews/pull/59))
 
 ## v0.4.0